Move websocket logic into Durable Object and out of the worker.
- simplifies update logic

- keeps all websocket management and update logic
in one central place

- pass gameId in initial websocket setup
