+++ workers/src/websocket-api.ts
@@ -1,4 +1,4 @@
-import { getCurrentUserFromCookie, getCurrentUserFromSessionId } from './auth';
+import { getCurrentUserFromCookie } from './auth';
 
 export async function handleSocket(request, env) {
   const user = await getCurrentUserFromCookie(request, env);
@@ -6,74 +6,15 @@ export async function handleSocket(request, env) {
     return new Response('Unauthorized', { status: 401 });
   }
 
-  const upgradeHeader = request.headers.get('Upgrade');
-  if (upgradeHeader !== 'websocket') {
-    return new Response('Expected websocket', { status: 400 });
+  let url = new URL(request.url);
+  let params = new URLSearchParams(url.search);
+  const gameId = params.get('gameId');
+  if (!gameId) {
+    return new Response('Missing gameId.', { status: 400 });
   }
 
-  const [client, server] = Object.values(new WebSocketPair());
+  const id = env.GAME_DO.idFromName(gameId);
+  const obj = env.GAME_DO.get(id);
 
-  server.accept();
-
-  server.addEventListener('close', () => {
-    console.log('websocket closed');
-  });
-
-  server.addEventListener('error', e => {
-    console.log('websocket error', e);
-  });
-
-  server.addEventListener('message', async event => {
-    server.send(
-      JSON.stringify({
-        eventId: 'debug',
-        eventData: 'message received for game' + event.data
-      })
-    );
-    const { sessionId, gameId, eventId, eventData } = JSON.parse(event.data);
-
-    // retrieve and verify user
-    const user = await getCurrentUserFromSessionId(sessionId, env);
-    if (!user) {
-      console.log('invalid user');
-      return;
-    }
-
-    // TODO: retrieve and verify gameId
-
-    const id = env.GAME_DO.idFromName(gameId);
-    const obj = env.GAME_DO.get(id);
-
-    // process message
-    switch (eventId) {
-      case 'vote':
-        const vote = eventData;
-        await obj.fetch(
-          new Request(
-            `http://durable/update?` +
-              new URLSearchParams({ vote, user: JSON.stringify(user) })
-          )
-        );
-        break;
-      case 'update-story':
-        const newStory = eventData;
-        await obj.fetch(
-          new Request(
-            `http://durable/update?` +
-              new URLSearchParams({
-                story: newStory,
-                user: JSON.stringify(user)
-              })
-          )
-        );
-        break;
-      default:
-        console.log('unknown websocket event', event.data);
-    }
-  });
-
-  return new Response(null, {
-    status: 101,
-    webSocket: client
-  });
+  return await obj.fetch(request);
 }
